# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TradeBox HYIP Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # SHIPMENTS
...
  /investments:
    get:
      operationId: listInvestments
      tags: [investments]
      summary: List my investments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, delivered, all]
      responses:
        "200":
          description: My investments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investment"
  /investments/{id}:
    get:
      operationId: getInvestment
      tags: [investments]
      summary: Get investment detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Investment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"

  # WALLET
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get full profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/kyc:
    post:
      operationId: submitKyc
      tags: [profile]
      summary: Submit KYC documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycInput"
      responses:
        "200":
          description: KYC submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/wallet-addresses:
    patch:
      operationId: updateWalletAddresses
      tags: [profile]
      summary: Update payout wallet addresses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletAddressesUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # MARKET
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        referralCode:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, traderId, balance, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        traderId:
          type: string
        balance:
          type: number
        role:
          type: string
          enum: [user, admin]
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        createdAt:
          type: string

    # SHIPMENT
    Shipment:
      type: object
      required: [id, title, cargoType, origin, destination, profitPercent, riskGrade, fundingGoal, fundingRaised, minInvestment, departureDate, arrivalDate, transitDays, status, freightForwarder, vesselName]
      properties:
        id:
          type: integer
        title:
          type: string
        cargoType:
          type: string
          enum: [electronics, cocoa, lithium, coffee, textiles, pharmaceuticals, cotton, steel]
        origin:
          type: string
        destination:
          type: string
        originCoords:
          type: ["string", "null"]
        destinationCoords:
          type: ["string", "null"]
        profitPercent:
          type: number
        riskGrade:
          type: string
          enum: [A, B, C, D]
        fundingGoal:
          type: number
        fundingRaised:
          type: number
        minInvestment:
          type: number
        departureDate:
          type: string
        arrivalDate:
          type: string
        transitDays:
          type: integer
        status:
          type: string
          enum: [open, funded, in_transit, delivered]
        freightForwarder:
          type: string
        vesselName:
          type: string
        createdAt:
          type: string
...
    Investment:
      type: object
      required: [id, shipmentId, amount, profitPercent, status, createdAt]
      properties:
        id:
          type: integer
        shipmentId:
          type: integer
        shipment:
          $ref: "#/components/schemas/Shipment"
        amount:
          type: number
        profitPercent:
          type: number
        expectedProfit:
          type: number
        actualProfit:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, delivered, cancelled]
        createdAt:
          type: string
        deliveredAt:
          type: ["string", "null"]

    # WALLET
...
    Profile:
      type: object
      required: [id, email, traderId, kycStatus, telegramHandle, whatsappNumber, twoFactorEnabled, walletAddresses, traderStats]
      properties:
        id:
          type: integer
        email:
          type: string
        traderId:
          type: string
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        telegramHandle:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        twoFactorEnabled:
          type: boolean
        walletAddresses:
          type: object
        traderStats:
          $ref: "#/components/schemas/TraderStats"
    TraderStats:
      type: object
      required: [totalShipped, totalProfit, countriesTraded, activeInvestments, deliveredInvestments]
      properties:
        totalShipped:
          type: number
        totalProfit:
          type: number
        countriesTraded:
          type: integer
        activeInvestments:
          type: integer
        deliveredInvestments:
          type: integer
    ProfileUpdate:
      type: object
      properties:
        telegramHandle:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        twoFactorEnabled:
          type: ["boolean", "null"]
    KycInput:
      type: object
      required: [idDocumentUrl, selfieUrl]
      properties:
        idDocumentUrl:
          type: string
        selfieUrl:
          type: string
        proofOfAddressUrl:
          type: ["string", "null"]
    WalletAddressesUpdate:
      type: object
      properties:
        btc:
          type: ["string", "null"]
        eth:
          type: ["string", "null"]
        usdt:
          type: ["string", "null"]
        bnb:
          type: ["string", "null"]

    # MARKET
...